Merged PR 15247: Update the test results sharing pipeline not to use MSSONIC-TOKEN
Update the test results sharing pipeline not to use MSSONIC-TOKEN

The classical service account sonicbld@microsoft.com will not be allowed in AzureDevOps. As a consequence, the PATs associated with this service account will no longer be valid:

MSSONIC-TOKEN
MSAZURE-TOKEN

All pipelines using these PATs must be updated to use compliant authentication methods.

This change updated the test results sharing pipeline to use the builtin System.AccessToken for authenticating with Azure DevOps.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>

----
#### AI description  (iteration 1)
#### PR Classification
This pull request is a pipeline configuration update.

#### PR Summary
This PR updates the test results sharing pipeline by removing the deprecated MSSONIC-TOKEN usage and substituting it with System.AccessToken to streamline authentication.
- `/.azure-pipelines/vendor_result_sharing/test_result_sharing.yml`: Removed the `sonicbld` group and replaced `$(MSSONIC-TOKEN)` with `$(System.AccessToken)`.
- `/.azure-pipelines/vendor_result_sharing/test_result_sharing_dualtor.yml`: Similar changes applied—removed the `sonicbld` group and updated the token variable from `$(MSSONIC-TOKEN)` to `$(System.AccessToken)`.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->