Merged PR 15322: Fix dualtor 7260 portchannel port down
Fix dualtor 7260 portchannel port down

The port down is due to the FEC mismatch between DUT and fanout; the DUT
FEC is disabled while the fanout is not. And the DUT FEC is disabled due
to `msft_an_enabled` is enabled.

Let's disable `msft_an_enabled` to unblock 7260 dualtor
qualification.

Signed-off-by: Longxiang Lyu <lolv@microsoft.com>