Fix qos sai test issue
When IPv6 is not enabled on the PTF, it sends RS packets with DSCP 0, which corresponds to PG 1. Consequently, if we test PG 0, the PG watermark is affected because some part of  the egress port buffer is occupied. Since the PTF sends RS packets randomly, the tests for `testQosSaiPgSharedWatermark[single_asic-wm_pg_shared_lossy]` and `testQosSaiQSharedWatermark[single_asic-wm_q_shared_lossy]` fail intermittently.