Fix retry count test case to properly teardown in case of failure (#20196)
This PR contains the following changes to the retry count test cases:

Use command instead of shell to avoid needing to start a shell where it's not needed.
Fix some Python style issues.
Rework the retry count test case logic to do a proper tear down by not having an assertion in a fixture. Instead, move the assertion to the test case itself. That makes sure that the teardown part of a fixture runs.
Fix the teamd kill test cases failing when running them individually in Elastictest, due to commands taking more time to run there.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>