Merged PR 15246: Update the sync-conn-graph pipeline not to use MSSONIC-TOKEN
Classical service account sonicbld@microsoft.com will not be allowed in AzureDevOps. As a consequence, the PATs associated with this service account will no longer be valid:
* MSSONIC-TOKEN
* MSAZURE-TOKEN

All pipelines using these PATs must be updated to use compliant authentication methods.

This change updated the sync conn graph pipeline to use the builtin System.AccessToken for authenticating with Azure DevOps.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>

----
#### AI description  (iteration 1)
#### PR Classification
This pull request is a pipeline configuration update.

#### PR Summary
The pull request replaces the use of `MSSONIC-TOKEN` with `System.AccessToken` in the sync connection graph pipeline for improved token management.
- `.azure-pipelines/repo_mgmt/sync_conn_graph.yml`: Updated the environment variable from `$(MSSONIC-TOKEN)` to `$(System.AccessToken)`.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->