Update expect_regex and ignore_regex for test_duplicate_route to support bulk update mode (#20232)
The sonic-sairedis change from sonic-net/sonic-sairedis#1613 modified the behavior of BULK operations which resulted in the log spew looking different before and after on API failures. This impacts test_duplicate_route.py. As this test is failing with the above sonic-sairedis change included, sonic-buildimage PRs that bump up the submodule reference fail pipeline checks.

The fix is to update the regexes so that the test can continue to pass with and without this sonic-sairedis change.

Summary:
Fixes # 20231

Signed-off-by: yenlu <yenlu@cisco.com>