Add transceiver status flag clear time to keys to skip comparison in scan_dbs (#19966)
test_add_rack currently checks teh value of key "TRANSCEIVER_STATUS_FLAG_CLEAR_TIME|Ethernetxxx" while the value of the key contains date and time. The key is for optical transaceivers so that when we have optical transceivers we will get state db comparison failure. We need to skip the comparison for these keys

Signed-off-by: Yael Tzur <ytzur@nvidia.com>