Fix the test_tunnel_decap_dscp_to_pg_mapping test case (#20313)
This is to fix a test issue only observed on MSFT testbed due to noisy
packets received by dut during the test. The size of the packet is very
big, so the base watermark of PG0 has an unexpected big initial value which
casues test failure. This initail value should be ignored.

Signed-off-by: opcoder0 <110003254+opcoder0@users.noreply.github.com>