Handle Multi-port portchannel member case in ECMP member flap TC (#20125)
This fixes issue mentioned in #17887.
For pizza box DUTs, multi-port portchannel scenario was not handled, which caused issue on certain specific topologies.

How did you do it?
Add all members of PORT_CHANNEL (if the port being shutdown is part of PORT_CHANNEL) to the filtered ports list, so that during fib_test, these ports are not randomly picked as src port.

Signed-off-by: opcoder0 <110003254+opcoder0@users.noreply.github.com>