Update calculation buffer in trimming case (#20225)
In the packet trimming cases, it needs to calculate the available buffer size of the egress queue and then block the queue to trigger the trimming packet.
There is a new feature "Update QoS to disable port alpha" (sonic-net/sonic-buildimage#23381). After the change, the calculation method of available buffer size has changed. So need update the buffer calculation function in trimming case

Signed-off-by: Yael Tzur <ytzur@nvidia.com>