Added new tests for scenarios where routes should disappear during a config reload.
Added static route tests to check for routes that routes don't exist after a config reload if not saved. Did the same for a warm reboot where we expect the unsaved routes to still be present. Also parameterized the file to make it shorter and easier to read the allure output.

Signed-off-by: Connor Roos <croos@nvidia.com>