Add fixture to provide designed routes information of topology file (#20238)
What is the motivation for this PR?
In some user scenario, for example bgp scale test in this PR, we need designed routes information of topology file to announcee routes on exabgp

How did you do it?
add a new fixture to provide designed routes information of topology file, and use the fixture in scale test to help annouce routes on exabgp.

How did you verify/test it?
run on testbed