[DHCP Relay counter] Add a test case for testing DHCP server sends packets to standby dut of dualtor (#20353)
What is the motivation for this PR?
At the point when a dualtor switch the active and standby host, the DHCP server may response the packets to standby host (active host previously). In this scenario, the standby should relay the packets and the client should get the relayed packet. We need a test case for it.

How did you do it?
Send offer, ack, nak and unknown packet to standby host and verify the packets on the client port.

How did you verify/test it?
run the test case dhcp_relay/test_dhcp_relay.py::test_dhcp_relay_on_dualtor_standby

Signed-off-by: Yael Tzur <ytzur@nvidia.com>