Issue 21164: [dhcp_relay] Add workaround for ACL limit issue
Summary:
Some tests which use ACLs can fail when all ACLs have been programmed.
We can detect this case and compensate for it by storing the table info
and re-instating it at the end of the test.

In the test infra, store the contents of the DATAACL table and restore
those contents at the end of the test.

Test fails without this change,  but passes with the change.