T2:snappi_tests:Change the speed_tol to correct number in pfcwd_actions. (#20074)
Description of PR
Summary:
The pfcwd_actions test: test_pfcwd_disable_pause_cngtn is being flaky due to setting the loss tolerance low. This is a regression introduced by #19772.

In this current PR, we are reverting the changes in traffic rate introduced by 19772 for this particular testcase.

Type of change
 Bug fix
 Testbed and Framework(new/improvement)
 New Test case
 Skipped for non-supported platforms
 Test case improvement

signed-off-by: jianquanye@microsoft.com
Signed-off-by: Guy Shemesh <gshemesh@nvidia.com>