Ignore pkt trim errors on non-TH5 ASICs (#20343)
What is the motivation for this PR?
Lots of new errors are popping out because of Packet TRIM feature enablement.

How did you do it?
Added a fixture in log analyzer to ignore these logs only for non-TH5 asics.

How did you verify/test it?
Verified manually on a non TH5 asic

Signed-off-by: yenlu <yenlu@cisco.com>