Ignore pkt trim errors on non-TH5 ASICs (#20465)
Description of PR
Summary:
Packet trim feature was recently introduced and it is supported only in
TH5 platforms. But the code has introduced lots of new ERR logs which
are resulting the test case to fail.

Though we are constantly working to fix the ERR logs, some are emitted
from SAI which has to be fixed from Broadcom(CSP12420291). Until then we can add
these error logs to be ignored.

Fixes #
https://github.com/aristanetworks/sonic-qual.msft/issues/757

Type of change
 Bug fix
 Testbed and Framework(new/improvement)
 New Test case
 Skipped for non-supported platforms
 Test case improvement

Approach
What is the motivation for this PR?
Lots of new errors are popping out because of Packet TRIM feature enablement.

How did you do it?
Added a fixture in log analyzer to ignore these logs only for non-TH5 asics.

How did you verify/test it?
Verified manually on a non TH5 asic

signed-off-by: jianquanye@microsoft.com