Add qos params for Arista-7060X6-16PE-384C-B-O128S2 (#20546)
Description of PR
This adds QoS parameters and the necessary test case changes to use them. This generalizes the handling of pkts_num_egr_mem to be applied to all Arista-7060X6* SKUs and updates its value for TH5-512.

Changes were manually tested and verified for 100% pass rate of the qos/test_qos_sai.py tests.

signed-off-by: jianquanye@microsoft.com
Signed-off-by: Guy Shemesh <gshemesh@nvidia.com>