Issue 21166: [dhcp_relay] Enable DHCP relay for T1 testbed tests
DHCP relay is not enabled for T1 topologies by default. Add a flag that
will optionally enable the feauture for the duration of the test.

Call `sudo config feature state dhcp_relay` with `enabled` before the
test and `disabled` afterwards (so that you leave the dut in the same
state you found it in).

Test is no longer skipped (it passes instead).