Fix dropcounter error when set before fdb config
There seems to be a bug on Moby where if a drop counter is added and
configured, followed by adding a static FDB entry, then the drop counter
would not be able to count packets properly.

This can be circumvented by creating the drop counter after the FDB entry
is added.

Note: moving setup_counters() would mean that the check to make sure the
device is capable of supporting the specific drop reasons will happen
after the FDB entry is added - which is in a try block and only cleaned
up in the finally block. However, it poses no issue as the finally block
will be executed even if the pytest skip is triggered in the try block.

Test no longer fails on Moby, ERR syslogs no longer gets triggered.