Disable log analyzer for k8s GCU test case to avoid flaky failure (#20639)
What is the motivation for this PR?
For this GCU test case, it doesn't need to enable log analyzer. And the log analyzer reports some error logs which are not relevant to this case itself, so need to disable the log analyzer to avoid flaky failure.

How did you do it?
Disable the log analyzer for this case.

How did you verify/test it?
Run it in a testbed, it passed.

Signed-off-by: Guy Shemesh <gshemesh@nvidia.com>