Issue 21169: [ptf] Use teamd for PTF portchannel
PTF containers use veth interfaces, which makes creating
portchannels with them tricky, since they lack the hardware information
needed for LACP negotiation. teamd works around this limitation, so
modify the test infra to detect when virtual interfaces are being added
to a LAG and use teamd instead of kernel bonding in that case.

Detect when virtual interfaces are being used as LAG pairs on the PTF
and switch to using `teamd` if they are. I left the original kernel
bonding code in place for backwards compatibility.

Manually verified that the LACP negotiation terminates correctly.
Verified that `snmp/test_snmp_fdb.py` passes with these changes.

Signed-off-by: Anders Linn <anders@nexthop.ai>