Skipping PFCWD tests for lossy topos & add anchor section (#20643)
This is adding a new section to the tests_mark_conditions.yaml file so that common
conditions can be defined once, and then re-used in the rest of this file.

This section currently only defines lossyTopos, and this change uses that throughout
the tests_mark_conditions.yaml file.

This is adding a new section to the tests_mark_conditions.yaml file so that common conditions can be defined once, and then re-used in the rest of this file.

This section currently only defines lossyTopos, and this change uses that throughout the tests_mark_conditions.yaml file.

This is a manual cherry-pick of
#20326
#20642

Signed-off-by: David Meggy <davidm@arista.com>