Fix flakyness in test_stress_arp
The test sets the dest ip as 1 greater than the PTF IP (link) and this new IP
happens to be one of our server IPs.

When the server port is standby and we trigger delete on a FAILED neighbor we
run into this known issue.

The actual problem is caused due to a duplicate delete request.
There is a race between orchagent and tunnelmgr which leads to this error.
This has been a known issue, refer sonic-net/sonic-swss#2579.

There are 3 options to solve this issue:
1. Ignore the error log, but this can mask other valid issues.
2. select an IPv6 address which doesnt belong to one of the SERVER IPs.
        This is still not fool proof, if any of the Servers ip neighbor was
learnt as FAILED or INCOMPLETE due to any other unforeseen reason, we will still
see this error during this test.
3. We enforce all the ports on dut to be in active state.

Option 3 looks the best and the same is implemented in this patch.

Verified on the testbed