validate and setup routes before test (#20717)
What is the motivation for this PR?
To avoid routes change on exabpg.

How did you do it?
Check if routes are identical with design before test, if not identical, restart exabgp and announce routes to make sure routes are identical with design.

How did you verify/test it?
Run code on testbed

Signed-off-by: yenlu <yenlu@cisco.com>