Add test case test_dhcp_monitor_checksum_validation to verify dhcp monitor IP/UDP checksum validation feature (#20655)
What is the motivation for this PR?
Currently, we don't have test case to check if the DHCP montior's cheksum validation works or not.

How did you do it?
Send invalid packets with incorrect IP/UDP checksum and check of the counter of Malformed

How did you verify/test it?
run the test case test_dhcp_monitor_checksum_validation