[Snappi][Broadcom-DNX]:New ECN testcases for Broadcom-DNX platforms (#18887)
Description of PR
The original ECN dequeue test case has Kmin and Kmax to 50k and 51k respectively. This leaves only a single packet gap between Kmin and Kmax values (with packet-size of 1024B), and hence difficult to test congestion testcases.

(1)
Based on discussion, we changed the test case setting of Kmin and Kmax to 800k and 2000k, Now the test will have different packet counts - namely 800, 1600, and 2400 packets of 1024B.

New test case is specifically meant for Broadcom-DNX platform and ensures that packets in range of Kmin-Kmax have the ECN marking probability less than or equal to Pmax.

Test also uses range of Pmax values to ensure that different ECN marking probability works for different Pmax values.

There is an associated README file also created to explain in more detail about the testcase.

(2)
Second testcase does not change the default Kmin, Kmax and Pmax parameters. With default ECN parameters, around 8500 packets (1024B) are sent and checked for ECN marking to adhere the ECN marking probability.

(3)
Restpy is used for capturing and analyzing the packets. Existing Snappi infrastructure allows to capture only 1024 packets of 1024B, hence, slicing needs to be enabled on IXIA to capture wider range of packets. Hence, use of restpy.

(4)
The new support for dynamic port selection is also incorporated.

Summary:
Fixes #12903

Type of change
 Bug fix
 Testbed and Framework(new/improvement)
 New Test case
 Skipped for non-supported platforms
 Test case improvement

Approach
What is the motivation for this PR?
How did you do it?
Based on the platform - Broadcom-DNX, the kmin, kmax values are set to 800k and 2000k respectively and the packets are checked with offset for congestion.

Original ECN dequeue test is pristine and has no changes whatsoever. We will need another PR maybe to exclude it for the test runs on Broadcom-DNX platforms.

How did you verify/test it?
Tested on 100 and 400G Broadcom-DNX platforms.

Any platform specific information?
Test is ONLY applicable for Broadcom-DNX platforms. It will be skipped for other platforms.

Supported testbed topology if it's a new test case?
T2 topology ONLY.

signed-off-by: jianquanye@microsoft.com
Signed-off-by: selldinesh <dinesh.sellappan@keysight.com>