Fix devutils and snmp pdu for docker-sonic-mgmt upgrade
The docker-sonic-mgmt will be upgraded to Ubuntu 24.04 based. The python packages will be upgraded together. Due to breaking changes in the packages, the devutils and snmp pdu controller need to be fixed to address the compability issues.

Major changes in this PR:
* Updated snmp_pdu_controllers.py to use pysnmp 7.1.21 APIs.
* Moved the current snmp_pdu_controllers.py based on pysnmp 4.x API to snmp_pdu_controllers_legacy.py.
* Updated the pdu_manager.py to import correct snmp pdu controller based on pysnmp versions.

Other changes:
* Improve the import in ssh_utils.py to make it compatible with old and new paramiko package.
* Add code to show traceback in task_runner.py to help troubleshooting.
* Improve the import in ansible_fixture.py to be compatible with old and new pytest_ansible.

All the changes are compatible with old and new docker-sonic-mgmt.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>