Verify docker-sonic-mgmt based on Ubuntu 24.04
A docker-sonic-mgmt based on Ubuntu 24.04 was built and pushed to sonicdev-microsoft.azurecr.io:443/docker-sonic-mgmt:24.04.0

This PR is to take advantange of the PR testing infrastructure to run all tests on KVM testbed using this new docker-sonic-mgmt to find out all incompability issues.

The setup-container.sh tool was updated to use this new docker-sonic-mgmt.

The PR test template was updated to continue PR testing when there is failure. So that all failures can be found out in one attempt.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>