Add pre-defined impact area to PR test template (#20835)
What is the motivation for this PR?
Sometimes we may need to use pipeline to run some PR test scripts, but we need different params for different topologies, and they are all defined in pr_test_template.yml. It would cost extra maintenance effort if it was defined in different files, so it's a better choice to open an entrance to allow self-defined impact area.

How did you do it?
Add pre-defined impact area to PR test template, so if we want to run some checks like regression binary search, we can pass a certain impact area and run PR test in a minor area, and do not need to care about any common or extra parameter change for each topo in PR test.
Add missing param back for baseline test, the parameter was missing in last change.

Signed-off-by: yenlu <yenlu@cisco.com>