[ipfwd] fix ipfwd tes (#21035)
In the ipfwd test, bgp container will be stopped and restarted. Currently, we wait for 15 second for bgp container fully initialize, which is sometimes not enough. Change the sleep to wait_until bgp up, and increase the wait time to 60 seconds.

What is the motivation for this PR?
Sleep 15 second is not enough to guarantee bgp container fully initialized

How did you do it?
Change sleep to wait_until

How did you verify/test it?
Run tests

Signed-off-by: Yael Tzur <ytzur@nvidia.com>