[fanout] Bugfix for SONiC 202505 fanout config template (#21037)
Current config template could generate a broken json format when 'marvell-teralynx' not in asic_type like below. There is an extra comma at the end of line hostname. This PR is for fixing the issue.

    "DEVICE_METADATA": {
        "localhost": {
            "default_pfcwd_status": "disable",
            "hwsku": "dummy-hwsku",
            "hostname": "dummy-hostname",
        }
    }

What is the motivation for this PR?
Current config template could generate a broken json format when 'marvell-teralynx' not in asic_type like below. There is an extra comma at the end of line hostname. This PR is for fixing the issue.

How did you do it?
Modify the sequence of type and hostname.

How did you verify/test it?
Verified by deploy an Arista fanout switch.