[arp test] Skip adding IP to interface in VLAN for topo type t0 (#21034)
The current code always tries to add an IP to the interface selected for testing. For topo type T0, usually an interface in VLAN is selected for testing. Adding IP to interface in VLAN wouldn't work. Before fix https://github.com/sonic-net/sonic-utilities/pull/3901, the failure is sliently ignored because the rc code is 0. This issue was not being noticed for quite some time. The tests using this fixture can pass anyway on topo type t0.

```python
    if (interface_is_in_vlan(vlan_member_table, interface_name)):
        click.echo("Interface {} is a member of vlan\nAborting!".format(interface_name))
        return
```
Here `click.echo` is used. The return code will be 0.

After fix https://github.com/sonic-net/sonic-utilities/pull/3901, adding IP to an interface in VLAN will explicitly fail with rc=2.
```python
    if (interface_is_in_vlan(vlan_member_table, base_interface_name)):
        ctx.fail("Interface {} is a member of vlan\nAborting!".format(base_interface_name))
        return
```

The tests using this fixture start to fail unnecessarily.

This PR fixed this problem by skipping adding IP to interface selected for testing if topo type is t0.

Signed-off-by: Xin Wang <xiwang5@microsoft.com>
Signed-off-by: Yael Tzur <ytzur@nvidia.com>