Add pre defined PR test template for regression binary search (#21017)
What is the motivation for this PR?
We are planning to add an automated project to auto detect and revert regression PRs, it would need a template to run a minor PR test with specific scope.

How did you do it?
Add pre defined PR test yml to allow users to run a minor PR test.
Modify get impacted area yml to allow variables be passing among steps.
Modify pr test template yml to add new and previous missing params.
Previously I created PR #20988 but got error when we do not pass IMPACT_AREA_INFO, because it was defined as object, if we do not pass it, it will be a null string and was not defined as an empty object, so the yml thought it was pre-defined.