set the protocol field to 'kernel' in generate_route_file (#21002)
With fib suppression enabled sonic is expected to ack with protocol field to zebra
However the protocol field is set as empty in the swssconfig loaded for every route. This when converted to int crashes. The protocl field is mandatory. Request to set protocol field in the test to avoid the log

Signed-off-by: selldinesh <dinesh.sellappan@keysight.com>