Fix ipfwd assert assumption for TH5 (#21072)
Summary: Fix ipfwd assert assumption for Broadcom TH5

The test limits the number of nhops that is used for testing. However,
the general assert assumes this limit does not exist and asserts with
CRM to make sure nhops available after the test is 0.

This causes issues on TH5 because the number of nhop groups supported
is 4095, where the limit of the test is 1024. This means that the number
of nhop groups after the test would not be 0 (4095 - 1024 - (extra groups) >> 0).

The Cisco-case of the test has the proper assert that adjusts for this
test limit. Fixing the test by including Broadcom TH5 devices to use that
proper assert.

Signed-off-by: opcoder0 <110003254+opcoder0@users.noreply.github.com>