Disable ipv6 traffic from ptf to  dut for two qos cases because it will introduce noise (#20416)
When IPv6 is enabled on the PTF, PTF will send RS packets with DSCP 0, which corresponds to PG 1. Consequently, if we test PG 0, the PG watermark is affected because some part of the egress port buffer is occupied for TC 1. Since the PTF sends RS packets randomly, the tests for testQosSaiPgSharedWatermark[single_asic-wm_pg_shared_lossy] and testQosSaiQSharedWatermark[single_asic-wm_q_shared_lossy] fail intermittently.
To stabilize the cases, we disable the ipv6 traffic from ptf to dut for the two cases.

Signed-off-by: Alberto Villarreal <alberto.villarreal@keysight.com>