Increase the pre_reboot_timeout for fwutil ONIE/BIOS test for SN4280
The switch shutdown time for SN4280 is longer due to
PR:https://github.com/sonic-net/sonic-buildimage/pull/24165. We need
increase the timeout.
Previousl we added this change only to 202506 branch, but now we decide
add it to master regradless of the status of sonic-buildimage PR#24165
status in master branch.
This is because the reboot time for SN4280 will be longer anyway due to
the DPUs.

Signed-off-by: Cong Hou <congh@nvidia.com>