add wait_until for TSC (#21047)
- Wrap the TSC call in a wait_until function
- Fix for https://github.com/sonic-net/sonic-mgmt/issues/20736
- Remove pytest_fail inside get_traffic_shift_state

What is the motivation for this PR?
Most test cases in the TSA/TSB suite check the Traffic shift state by issuing TSC immediately after a TSA/TSB operation. However, due to some delay in the application of FRR config, TSC can report that the DUT is in inconsistent state if we read the config before it is fully applied, leading to intermittent test failures.

How did you do it?
I've modified the code to add retry mechanism to the TSC check.

How did you verify/test it?
Ran the TSA/TSB tests on a T2 topology.

Signed-off-by: Alberto Villarreal <alberto.villarreal@keysight.com>