[ansible]: Set sysctl args and ptf memory for full topology (#21097)
What is the motivation for this PR?
We want to add a full topo with 448 neighbors; the previous configuration and memory limitation cannot meet its requirement.

How did you do it?
Increase the memory limitation of PTF and increase the max number of inotify user instances.

How did you verify/test it?
Check it locally

Signed-off-by: opcoder0 <110003254+opcoder0@users.noreply.github.com>