[Test improvement] Make WrrTest more robust to packet loss at ptf
Description of PR
Summary:

The test_qos_sai::testQosSaiDwrr and testQosSaiDwrrWeightChange validates the packets received from the dut at the ptf. Sometimes when the packets are sent from high speed interfaces to ptf in a burst, packets get lost. The test should handle packet loss and still do validation instead of asserting 100% packets got received.

Fixes # (issue)

Type of change
Test case improvement

Approach
What is the motivation for this PR?
The test_qos_sai::testQosSaiDwrr and testQosSaiDwrrWeightChange fails even if only a handful of packets get dropped the ptf.

How did you do it?
Added a parameter scale_factor (default: 80%), can be tuned to be higher in the yaml files. Adjust the number of packets expected and margin or error based on this scale factor and do the same validation

How did you verify/test it?
Ran the test successfully even with packet drops. Used 0.95 scale factor in the yaml